Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-    int num1 , num2 , num3 ;
+    float num1 , num2 , num3 ;
     float av;
     scanf("%i %i %i" , &num1 , &num2 , &num3);
     av = ((num1+num2+num3)/3);
